modified form with checkbox
