+++ client/components/SingleUser.js
@@ -5,7 +5,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { fetchUser } from "../redux/users";
 import CompletedTrips from "./CompletedTrips";
 import ActiveTrips from "./ActiveTrips";
-import AllTasks from "./AllTasks";
+import AllTasks from "./TaskComponents/AllTasks/AllTasks";
 
 const SingleUser = () => {
   const dispatch = useDispatch();
